Code2.0/requirements.txt
fastapi>=0.100.0
uvicorn[standard]>=0.23.0
python-multipart>=0.0.6
pdfplumber>=0.10.0
torch>=2.6.0
torchvision
torchaudio
langchain>=0.1.0
langchain-community>=0.0.10
langchain-openai>=0.0.5
langchain-chroma>=0.1.0
chromadb>=0.4.22
pydantic>=2.0.0
pydantic-settings>=2.0.0
sqlalchemy>=2.0.0
sentence-transformers>=2.2.2
tiktoken>=0.5.0
python-pptx>=0.6.21
python-docx>=0.8.11
cairosvg>=2.7.1
